refactor(marketing): extract the cadence handler and record the posture change (BI-C26FE785)
Preflight guard fixes for the marketing-campaign proactivity family.

Module Size Guard and the substrate-complexity ratchet both failed for the
same reason: adding the posture-aware handler pushed marketing-ops-pack.ts to
817 LOC against an 800 ceiling, taking the production hotspot count 65 -> 66.
Extracted planUpcomingMarketingDraftsHandler into marketing-cadence-handler.ts,
following the existing one-handler-per-sibling shape of
backlog-update-item-handler.ts. The pack is back to 773 LOC and the hotspot
count is unchanged. The baseline was not loosened.

Docs Impact Gate flagged apps/web/lib/marketing/scheduler.ts against the
long-running-agentic-process architecture doc, correctly: the scheduler's lead
time stopped being a constant. Updated section 7.2 (Marketing campaigns) to
record that the brief -> draft transition is now a governed decision rather
than a fixed ADVANCE_DAYS_FOR_DUE_WINDOW, and that the section's own
non-bypassable HITL boundary is now enforced twice — once at the approval
queue and once at the resolver, which caps the family at "propose" for every
level.

No behaviour change: 45 test files / 361 tests still pass after the extraction.

Design-Grounding-Decision: No contract change. This is a mechanical extraction to satisfy the module-size ratchet plus a doc update recording behaviour that landed in the parent commit. Source of truth remains docs/superpowers/specs/2026-05-26-marketing-execution-loop-design.md; the architecture doc section 7.2 is a downstream description of the same seam and was brought back in step with the code.

Docs-Impact-Decision: Updated docs/architecture/2026-06-09-long-running-agentic-process-architecture.md section 7.2 rather than attesting the change away — the scheduler's cadence genuinely changed from a constant to a posture-resolved value, which that section describes.

Process-Spine-Decision: New module marketing-cadence-handler.ts is an extraction of existing committed code to satisfy the module-size ceiling, not a new capability; the capability it serves is documented in the architecture doc update included in this commit.
